Rename views.h to view_containers.h, commpile fixes
