+++ src/kv/store.h
@@ -5,7 +5,7 @@
 #include "kv_serialiser.h"
 #include "kv_types.h"
 #include "map.h"
-#include "views.h"
+#include "view_containers.h"
 
 namespace kv
 {
